+++ src/app/auth/finalize/layout.tsx
@@ -1,0 +1,5 @@
+import { Suspense } from 'react';
+
+export default function Layout({ children }: { children: React.ReactNode }) {
+	return <Suspense>{children}</Suspense>;
+}
