--- src/app/login/[provider]/layout.tsx
@@ -1,5 +1,0 @@
-import { Suspense } from 'react';
-
-export default function WithoutSideLayout({ children }: { children: React.ReactNode }) {
-	return <Suspense>{children}</Suspense>;
-}
